refactor to model package, fix min/max temp associations
